requirements.txt
ccxt
pandas
numpy
scikit-learn
matplotlib
joblib
pyyaml
logging